requirements.txt
asgiref==3.2.3
bcrypt==3.1.7
cffi==1.14.0
Django==3.0.3
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
Markdown==3.2.1
mysqlclient==1.4.6
pycparser==2.20
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
requests==2.23.0